Improvement: Running even more in the background
This means most of the time you can use LaunchBar again immediately.
Only occasion where it is not running in the background is if there are no grouped notifications and are using it dismiss these. In this case I would rather use the action that is meant just for that use case.