Audio Devices 0.2
Still beta. Tried to include AirPlay devices. But it is tricky. CoreAudio does not list AirPlay devices consistently. You can try to add or update AirPlay devices by holding `Shift`, but this is a workaround that involves GUI scripting, which is very fragile. I might remove that whole part again.