Zotero Paste Helper 1.0
Zotero Paste Helper is a companion action to Zotero Search V2 written in Swift that is not required but largely improves pasting rich text. It includes a fallback to Markdown for apps that can’t handle rich text. It needs to be compiled to run properly. You can use "Compile Swift Action" to do so.